Simplified API functionality. Split createContainer into creation and stream linking parts. Also made separate test file.
